+++ test/js/bun/http/bun-server.test.ts
@@ -183,16 +183,18 @@ describe.concurrent("Server", () => {
 
   test("abort signal on server", async () => {
     {
-      let signalOnServer = false;
+      let abortPromise = Promise.withResolvers();
+      let responseAwaited = Promise.withResolvers();
       let fetchAborted = false;
       const abortController = new AbortController();
       using server = Bun.serve({
         async fetch(req) {
           req.signal.addEventListener("abort", () => {
-            signalOnServer = true;
+            abortPromise.resolve();
           });
           abortController.abort();
           await Bun.sleep(15);
+          responseAwaited.resolve();
           return new Response("Hello");
         },
         port: 0,
@@ -206,8 +208,7 @@ describe.concurrent("Server", () => {
         fetchAborted = true;
       }
       // wait for the server to process the abort signal, fetch may throw before the server processes the signal
-      await Bun.sleep(15);
-      expect(signalOnServer).toBe(true);
+      await Promise.all([abortPromise.promise, responseAwaited.promise]);
       expect(fetchAborted).toBe(true);
     }
   });
