+++ test/js/web/fetch/fetch.tls.test.ts
@@ -260,35 +260,6 @@ describe.concurrent("fetch-tls", () => {
       expect(total).toBeLessThanOrEqual(TIMEOUT + THRESHOLD);
     }
   });
-  for (const timeout of [0, 1, 10, 20, 100, 300]) {
-    it(`fetch should abort as soon as possible under tls using AbortSignal.timeout(${timeout})`, async () => {
-      using server = Bun.serve({
-        port: 0,
-        tls: CERT_LOCALHOST_IP,
-        async fetch() {
-          await Bun.sleep(1000);
-          return new Response("Hello World");
-        },
-      });
-      const THRESHOLD = 50;
-
-      const time = performance.now();
-      try {
-        await fetch(server.url, {
-          //@ts-ignore
-          tls: { ca: CERT_LOCALHOST_IP.cert },
-          signal: AbortSignal.timeout(timeout),
-        }).then(res => res.text());
-        expect.unreachable();
-      } catch (err) {
-        expect((err as Error).name).toBe("TimeoutError");
-      } finally {
-        const diff = performance.now() - time;
-        expect(diff).toBeLessThanOrEqual(timeout + THRESHOLD);
-        expect(diff).toBeGreaterThanOrEqual(timeout - THRESHOLD);
-      }
-    });
-  }
 
   it("fetch should use NODE_EXTRA_CA_CERTS", async () => {
     using server = Bun.serve({
